#227 - fixed problem with remove of the Open Graph fallback image

+++ app/back-end/themes.js
@@ -420,6 +420,7 @@ class Themes {
         let authorsData = authors.load();
         let authorAvatars = [];
         let ampFallbackImage = '';
+        let ogFallbackImage = '';
 
         if(authorsData && authorsData.length) {
             for(let i = 0; i < authorsData.length; i++) {
@@ -443,6 +444,11 @@ class Themes {
         if(siteData && siteData.advanced && siteData.advanced.ampImage) {
             ampFallbackImage = siteData.advanced.ampImage;
         }
+
+        if(siteData && siteData.advanced && siteData.advanced.openGraphImage) {
+            ogFallbackImage = siteData.advanced.openGraphImage;
+        }
+
         // Make sure that all slashes are in the same direction
         configString = normalizePath(configString);
 
@@ -469,7 +475,8 @@ class Themes {
             if(
                 configString.indexOf('/' + imagePath) === -1 &&
                 authorAvatars.indexOf(imagePath) === -1 &&
-                imagePath !== ampFallbackImage
+                imagePath !== ampFallbackImage &&
+                imagePath !== ogFallbackImage
             ) {
                 try {
                     fs.unlinkSync(fullPath);
